# [WIP] Initial attempt at faster slack message parser