As discussed in #32. It turns out the slack message format we're trying to parse isn't a very optimal use case for instaparse. 

Here's a first attempt at a parser that only makes use of regular expressions. In particular, this PR adds a creatively named `parse2` function to clojurians-log.message-parser. It mimicks the output of `parse` and should mostly be a drop-in replacement. At least it can pass all the same tests that `parse` passes.

---

Initial comparison indicates the new parser is roughly 100x faster than the old one. You can try this for yourself using the commented forms at the bottom of message_parser.clj.

On my machine, running `parse` over all messages on the "clojure" channel on 2018-02-02 took ~445ms. The `parse2` version took 4.3ms.

The routes where parse performance was initially discovered also completes in a much more reasonable time.

| route | time |
|------|------|
/clojure/2018-02-02 | 46.9 ms |
/clojure/2018-02-02/1517583973.000174 | 48.8 ms |

---

Tasks remaining:
- [ ] Integrate with existing user/channel reference parsing code (inline names already extracted)
- [ ] A lot more tests(?)

---

- [x] My code conforms to this project's [Style Guide](https://github.com/clojureverse/clojurians-log-app/blob/master/docs/STYLE.md)
- [x] I have added tests for functions or features I've added
- [x] All tests are green (`lein test`)