@Odie 2018-03-21
Looks like there are input cases where both `parse` and `parse2` fails.

For example:
```_*hello*_``` results in text that is both bold and italicized on slack.
The same input, when placed through either functions, results in: ``` [[:italic "hello"]] ```

Styled text also combines "interesting" ways, which also works here on github, like so:
~hello ```world```~