@plexus 2018-03-21
I appreciate the effort here, and it's impressive the numbers you're
getting, but I feel you're taking a very non-standard approach to parsing
that is really going to limit us.

Markdown is a "context free" language, regexes on their own aren't powerful
enough to represent that, which is why you're having trouble dealing with
nested syntax.

If we want to do this right and replace the instaparse version I think
it'll have to be an actual hand rolled parser like recursive-descent or
shift-reduce.

On Wed, Mar 21, 2018, 09:49 Jonathan Shieh <notifications@github.com> wrote:

> Looks like there are input cases where both parse and parse2 fails.
>
> For example:
> _*hello*_ results in text that is both bold and italicized on slack.
> The same input, when placed through either functions, results in: [[:italic
> "hello"]]
>
> Styled text also combines "interesting" ways, which also works here on
> github, like so:
> hello world
>
> —
> You are receiving this because your review was requested.
> Reply to this email directly, view it on GitHub
> <https://github.com/clojureverse/clojurians-log-app/pull/34#issuecomment-374813603>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAB91ND4rEY7uxa6XU5UOw2kO38Y3LU2ks5tgbGKgaJpZM4Svx2g>
> .
>