@jumarko 2018-03-21
You may want to look at https://github.com/aphyr/clj-antlr. It generally performs much better than instaparse.