@Odie 2018-03-21
Well, it’s been a long time since my college days and I’m afraid I don’t remember the finer points of context free languages. Though a quick google suggests that markdown is not context free. [link](https://news.ycombinator.com/item?id=13871781)

Anyway, I do think it’s more than possible to amend this current strategy to deal with nested structures. The regex results identifies interesting, potentially nested, regions. With the help of a stack, we should be able to 1) delay conversion of match data to “token” until we’re sure there are no more nested structures and 2) properly construct the nested structure with the help of the same stack. This is not unlike construction of the ast after a tolenization pass in a compiler.

I hope to have a version that implements this working sometime soon (tomorrow?). If you’re to be away for a bit, it ought to be ready when you’re back for sure. :)

Have a safe trip!