@plexus 2018-03-21
I played around myself a bit just now. I don't think my approach is going to be fruitful though, but I'll just leave these here :)

``` clojure
;; Try out Slack message parsing at
;; https://api.slack.com/docs/messages/builder?msg=%7B%22text%22%3A%22xx1_%20*basic*%60%22%7D

(deftest parse-test
  (are [x y] (= y (parse x))
    "basic"              ["basic"]
    "*bold*"             [[:bold "bold"]]
    "basic *bold* basic" ["basic " [:bold "bold"] " basic"]
    "basic *basic"       ["basic *basic"]
    "_italic_"           [[:italic "italic"]]
    "xx _italic_ xx"     ["xx " [:italic "italic"] " xx"]
    "xx_oops_"           ["xx_oops_"]
    "xx1`*basic*`"       ["xx1`*basic*`"]
    "xx1_`*basic*`"      ["xx1_" [:inline-code "*basic*"]]
    "xx1_ *basic*`"      ["xx1_ *basic*`"]
    "> foo *_bar_*"      [[:blockquote "foo " [:bold [:italic "bar"]]]]
    ">foo *_bar_*"       [[:blockquote "foo " [:bold [:italic "bar"]]]]
    ">>foo *_bar_*"      [[:blockquote ">foo " [:bold [:italic "bar"]]]]
    "> >foo *_bar_*"     ["> >foo " [:bold [:italic "bar"]]]]))
```

Note that I parsed "undecorated" to plain strings, but for the rest the structure should be about the same.

> You may want to look at https://github.com/aphyr/clj-antlr.

This seems like something we should seriously consider. Parsers seem fun to write until you're halfway through, getting all the details right gets extremely tedious.

> Though a quick google suggests that markdown is not context free

If it's not then it's context-sensitive, which is even harder to deal with. It's certainly not regular.