@Odie 2018-03-22
Well, the new code uses a stack to deal with nested regions properly. It's honestly not a super readable piece of code, but, at least it's passing all previous tests.

I'm also not sure how far I can push the regexes to accurately deal with every edge case to replicate the exact behavior of slack's parser.  Some of the test cases added with https://github.com/clojureverse/clojurians-log-app/pull/34/commits/f83d534b9d39349cb3178c10ec68b2ef5618dfbb was a bit hard to deal with. Will try sticking with this for a little while longer.

In the meantime, if anybody wants to give clj-antlr a shot, please do go ahead!