@Odie 2018-03-22
Quick note, the revised parser now takes ~7-9ms on 
```
  (let [data (clojurians-log.db.queries/channel-day-messages (user/db) "clojure" "2018-02-02")]
    (time (->> data
               (map #(parse2 (:message/text %)))
               doall))
    nil)
```
instead of the previous ~4ms.