@Odie 2018-03-23
Blockquote support added.

https://github.com/clojureverse/clojurians-log-app/pull/34/commits/09278a61c60013afe772d1119b9627f3a1d98fc4 also adds some performance fixes. This gives us a bit of a speed up, from 8.3ms => 6.9ms, measured with criterium.